transferred contexts to context folder
