+++ frontend/src/app/providers.tsx
@@ -1,7 +1,7 @@
-import { ProjectSidebarProvider } from "@/contexts/ProjectSidebarContext";
-import { UserSidebarProvider } from "@/contexts/UserSidebarContext";
+import { ProjectSidebarProvider } from "@/context/ProjectSidebarContext";
 import { ThemeProvider } from "@/context/ThemeContext";
 import { UserProvider } from "@/context/UserContext";
+import { UserSidebarProvider } from "@/context/UserSidebarContext";
 
 function Providers({ children }: { children: React.ReactNode }) {
   return (
