+++ frontend/src/components/user/UserSidebar.tsx
@@ -1,11 +1,12 @@
 "use client";
-import { useUserSidebar } from "@/hooks/useUserSidebar";
+
+import { useUserSidebarContext } from "@/context/UserSidebarContext";
 
 function UserSidebar() {
-  const { userSidebarRef, isUserSidebarOpen } = useUserSidebar();
+  const { userSidebarRef, isUserSidebarOpen } = useUserSidebarContext();
   return (
     <aside
-      className={`z-50 flex h-full w-[14rem] lg:w-[18rem] flex-col gap-10 px-[3.2rem] py-[2.4rem] text-3xl shadow-md ${isUserSidebarOpen ? "translate-x-0" : "-translate-x-full"}`}
+      className={`z-50 flex h-full w-[14rem] flex-col gap-10 px-[3.2rem] py-[2.4rem] text-3xl shadow-md lg:w-[18rem] ${isUserSidebarOpen ? "translate-x-0" : "-translate-x-full"}`}
       ref={userSidebarRef}
     >
       This is the User Sidebar component
