+++ frontend/src/context/ProjectSidebarContext.tsx
@@ -1,6 +1,6 @@
 "use client";
 import { useOutsideClick } from "@/hooks/useOutsideClick";
-import { createContext, useState, useRef } from "react";
+import { createContext, useState, useRef, useContext } from "react";
 import type { ProjectSidebarContextTypes } from "@/types/Project";
 
 const defaultContextValue: ProjectSidebarContextTypes = {
@@ -43,4 +43,17 @@ function ProjectSidebarProvider({ children }: { children: React.ReactNode }) {
   );
 }
 
-export { ProjectSidebarContext, ProjectSidebarProvider };
+
+
+function useProjectSidebarContext() {
+  const context = useContext(ProjectSidebarContext);
+  if (context === undefined) {
+    throw new Error(
+      "useProjectSidebar must be used within a ProjectSidebarProvider",
+    );
+  }
+  return context;
+}
+
+
+export { useProjectSidebarContext, ProjectSidebarProvider };
