--- frontend/src/hooks/useProjectSidebar.ts
@@ -1,12 +1,0 @@
-import { ProjectSidebarContext } from "@/contexts/ProjectSidebarContext";
-import { useContext } from "react";
-
-export function useProjectSidebar() {
-  const context = useContext(ProjectSidebarContext);
-  if (context === undefined) {
-    throw new Error(
-      "useProjectSidebar must be used within a ProjectSidebarProvider",
-    );
-  }
-  return context;
-}
